+++ kgml-core/src/commonTest/kotlin/dev/karmakrafts/kgml/transform/Matrix3x3TransformTest.kt
@@ -27,8 +27,8 @@ class Matrix3x3TransformTest {
         val rad = (PI / 2).toFloat()
 
         // Z-axis rotation
-        assertMatrix3x3Equals(Matrix3x3f.rotationZRad(rad), Matrix3x3f.rotationZ(90f))
-        val mZ = Matrix3x3f.rotationZ(90f)
+        assertMatrix3x3Equals(Matrix3x3f.rotationRad(angleZ = rad), Matrix3x3f.rotation(angleZ = 90f))
+        val mZ = Matrix3x3f.rotation(angleZ = 90f)
         assertEquals(0f, mZ.m00, 1e-6f)
         assertEquals(-1f, mZ.m01, 1e-6f)
         assertEquals(0f, mZ.m02, 1e-6f)
@@ -40,8 +40,8 @@ class Matrix3x3TransformTest {
         assertEquals(1f, mZ.m22, 1e-6f)
 
         // X-axis rotation
-        assertMatrix3x3Equals(Matrix3x3f.rotationXRad(rad), Matrix3x3f.rotationX(90f))
-        val mX = Matrix3x3f.rotationX(90f)
+        assertMatrix3x3Equals(Matrix3x3f.rotationRad(angleX = rad), Matrix3x3f.rotation(angleX = 90f))
+        val mX = Matrix3x3f.rotation(angleX = 90f)
         assertEquals(1f, mX.m00, 1e-6f)
         assertEquals(0f, mX.m11, 1e-6f)
         assertEquals(-1f, mX.m12, 1e-6f)
@@ -49,8 +49,8 @@ class Matrix3x3TransformTest {
         assertEquals(0f, mX.m22, 1e-6f)
 
         // Y-axis rotation
-        assertMatrix3x3Equals(Matrix3x3f.rotationYRad(rad), Matrix3x3f.rotationY(90f))
-        val mY = Matrix3x3f.rotationY(90f)
+        assertMatrix3x3Equals(Matrix3x3f.rotationRad(angleY = rad), Matrix3x3f.rotation(angleY = 90f))
+        val mY = Matrix3x3f.rotation(angleY = 90f)
         assertEquals(0f, mY.m00, 1e-6f)
         assertEquals(1f, mY.m02, 1e-6f)
         assertEquals(-1f, mY.m20, 1e-6f)
@@ -58,11 +58,13 @@ class Matrix3x3TransformTest {
 
         // Combined rotation
         val mXYZRad = Matrix3x3f.rotationRad(rad, rad, rad)
-        val expectedRad = Matrix3x3f.rotationXRad(rad) * Matrix3x3f.rotationYRad(rad) * Matrix3x3f.rotationZRad(rad)
+        val expectedRad =
+            Matrix3x3f.rotationRad(angleX = rad) * Matrix3x3f.rotationRad(angleY = rad) * Matrix3x3f.rotationRad(angleZ = rad)
         assertMatrix3x3Equals(expectedRad, mXYZRad)
 
         val mXYZ = Matrix3x3f.rotation(90f, 90f, 90f)
-        val expected = Matrix3x3f.rotationX(90f) * Matrix3x3f.rotationY(90f) * Matrix3x3f.rotationZ(90f)
+        val expected =
+            Matrix3x3f.rotation(angleX = 90f) * Matrix3x3f.rotation(angleY = 90f) * Matrix3x3f.rotation(angleZ = 90f)
         assertMatrix3x3Equals(expected, mXYZ)
     }
 
@@ -84,7 +86,7 @@ class Matrix3x3TransformTest {
     fun `Matrix3x3 rotation from quaternion should work`() {
         val q = Quaternion.fromAngles(90f, 0f, 0f)
         val m = Matrix3x3f.identity * q
-        assertMatrix3x3Equals(Matrix3x3f.rotationX(90f), m)
+        assertMatrix3x3Equals(Matrix3x3f.rotation(angleX = 90f), m)
     }
 
     private fun assertMatrix3x3Equals(expected: Matrix3x3f, actual: Matrix3x3f) {